+++ packages/expo-router/src/link/useRouter.ts
@@ -3,7 +3,7 @@ import { useCallback } from "react";
 import { Href, resolveHref } from "./href";
 import { useLinkToPath } from "./useLinkToPath";
 import { useLoadedNavigation } from "./useLoadedNavigation";
-import { RootContainer } from "../ContextNavigationContainer";
+import { getNavigationContainerRef } from "../NavigationContainer";
 
 // Wraps useLinkTo to provide an API which is similar to the Link component.
 export function useLink() {
@@ -23,7 +23,7 @@ type Router = {
 };
 
 export function useRouter(): Router {
-  const root = RootContainer.useRef();
+  const root = getNavigationContainerRef();
 
   const pending = useLoadedNavigation();
   const linkTo = useLinkToPath();
